ai-realtime/requirements.txt
pipecat-ai[daily,openai,silero]==0.0.106
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
aiohttp>=3.11.12,<4
